[add] ore subdir and init MiningDrillBlock
Basically, I added the foundation for the ore stuff, it all still needs
to be implemented and refined, but the general skeleton is there
following similar patterns to basin. It still needs functions, and other
such things, but I at least thing I have an understanding of how it
functions now and can be done I think with relative ease.

I touched the miningDrillBlock a tiny bit, but I was still unfamiliar so
while the changes are relevant, they might not be as good as the ore
stuff.

I primarily focused on learning how Create does stuff, so there isn't
much to show yet but it is a start.